Reorganize some extension functions
