+++ app/src/main/kotlin/com/simplemobiletools/musicplayer/fragments/MyViewPagerFragment.kt
@@ -21,7 +21,7 @@ abstract class MyViewPagerFragment(context: Context, attributeSet: AttributeSet)
 
     abstract fun setupColors(textColor: Int, adjustedPrimaryColor: Int)
 
-    fun prepareAndPlay(tracks: List<Track>, startIndex: Int = 0, startPosition: Long = 0, startActivity: Boolean = true) {
-        (context as SimpleControllerActivity).prepareAndPlay(tracks, startIndex, startPosition, startActivity)
+    fun prepareAndPlay(tracks: List<Track>, startIndex: Int = 0, startPositionMs: Long = 0, startActivity: Boolean = true) {
+        (context as SimpleControllerActivity).prepareAndPlay(tracks, startIndex, startPositionMs, startActivity)
     }
 }
